extopts: Basic string argument type is protected by const.
Now handling of string parameters is handled thoutgh 'const_str'
instead of 'addr' field of extopt structure. So EXTOPT_ARG_STR awaits
now 'const char **' instead of 'char **'.

This protects from using STR type for reading string argument and than
changing it while it is still located at command line arguments
string.

If changing is needed then STR_ALLOC type should be used or STR but
the contents of it must be copied to other location.

BE AWARE not to use 'addr' field for this type of arguments since in
this case it will not help you to track down if you accidentally
defined string as non-const and than also accidentally writing into
it. It is better to simply do not show off and use EXTOPT_ARG_STR
macro :3